Convert to inline resources and test create_if_missing
Inline resources greatly simplify the notification issues. This also
fixes requiring a command when you delete a cron entry, which was a
regression in 1.7.1. I've added a serverspec to ensure we're handling
create_if_missing correctly by first creating a cron entry, then
creating it again with create_if_missing.  We can expect to find the
content of the first creation only.