Put validator methods above where they are used.
I tried to put the validators into a seperate module, but I could not get
Chef to load them properly. I would prefer to do that since it would be
a cleaner approach, so if someone knows how to make that happen, I would
do that instead.

Fixes chef-cookbooks/cron#71